[ES] Implemented scroll-based query backend
Previously, there was a hard limit of 10,000 features that could be
returned by any single tile request. This limitation is imposed by
ElasticSearch so that it doesn't need to spend so much time during the
document fetching phase of a search request.

This commit removes this limitation by using the ElasticSearch Scroll
API, which allows the tilenol backend to basically scroll through 100%
of the search results from a query, and re-encode them into feature
collections.