Adds "-s" toggle to enable shape simplification (defaults to false)
In running performance tests, I've found that contrary to the suggestions made in #3,
the better solution in most cases is to not do shape simplification at all, which
reduces CPU loads at the cost of network payload size (since the result is technically
higher-fidelity vector geometries). That said, this has the practical advantage of
reducing latency on higher-bandwidth network connections, by relying more heavily on
network optimizations (caching, compression, etc.).